requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.6.0; python_version >= '3.7'
certifi==2022.12.7; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0; python_version >= '3.7'
cryptography==40.0.2; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-templated-mail==1.1.1
django==4.2
djangorestframework-simplejwt==5.2.2; python_version >= '3.7'
djangorestframework==3.14.0
djoser==2.2.0
idna==3.4; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
pillow==9.5.0
pycparser==2.21
pyjwt==2.6.0; python_version >= '3.7'
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.29.0; python_version >= '3.7'
social-auth-app-django==5.2.0; python_version >= '3.7'
social-auth-core==4.4.2; python_version >= '3.6'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2023.3; sys_platform == 'win32'
urllib3==1.26.15; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
